Husky alpha v.01 - it might actually run this time
I think that nothing is broken now. Even if it doesn't make any good trades, it trades properly according to the rules I set out. Next on the drawing board is what to do when someone beats me to the punch and snaps up the orders. I'm left with a large balance of altcoins and the ask I put out will be too high to be filled anytime soon. There is an opportunity cost to sitting on these open orders as well as the fact that these orders signal a decrease in value for the coin since we bought in. That means we bought something that is going down.